requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-google-genai
langchain-community
langchain-ollama
langgraph
chromadb
sentence-transformers
PyPDF2
pandas
python-dotenv
